// src/app/Models/SignosVitales.ts
export interface SignosVitales {
    Idper:                number;
    Idevento:             number;
    Fecha:                Date;
    Hora:                 string;
    Peso:                 number;
    Talla:                number;
    Imc:                  number;
    FrecuenciaCardiaca:   number;
    FrecuenciaSistolica:  number;
    FrecuenciaDiastolica: number;
    PerimetroCefalico:    number;
    PerimetroCintura:     number;
    SuperficieCorporal:   number;
    Temperatura:          number;
}